# Clear sensitive memory without getting optimized out.