--- a/src/bench_internal.c
+++ b/src/bench_internal.c
@@ -3,8 +3,6 @@
  * Distributed under the MIT software license, see the accompanying   *
  * file COPYING or http://www.opensource.org/licenses/mit-license.php.*
  **********************************************************************/
-#include <stdio.h>
-
 #include "include/secp256k1.h"
 
 #include "util.h"