--- a/src/ecmult.h
+++ b/src/ecmult.h
@@ -23,7 +23,6 @@ static void secp256k1_ecmult_context_build(secp256k1_ecmult_context *ctx, const 
 static void secp256k1_ecmult_context_clone(secp256k1_ecmult_context *dst,
                                            const secp256k1_ecmult_context *src, const secp256k1_callback *cb);
 static void secp256k1_ecmult_context_teardown(secp256k1_ecmult_context *ctx);
-static void secp256k1_ecmult_context_clear(secp256k1_ecmult_context *ctx);
 static int secp256k1_ecmult_context_is_built(const secp256k1_ecmult_context *ctx);
 
 /** Double multiply: R = na*A + ng*G */