--- a/src/ecmult_gen_impl.h
+++ b/src/ecmult_gen_impl.h
@@ -151,7 +151,7 @@ static void secp256k1_ecmult_gen(const secp256k1_ecmult_gen_context *ctx, secp25
         secp256k1_gej_add_ge(r, r, &add);
     }
     bits = 0;
-    secp256k1_ge_clear(&add);
+    SECP256K1_CLEANSE(add);
     SECP256K1_CLEANSE(gnb);
 }