--- a/src/field_5x52_impl.h
+++ b/src/field_5x52_impl.h
@@ -253,11 +253,22 @@ SECP256K1_INLINE static int secp256k1_fe_is_odd(const secp256k1_fe *a) {
     return a->n[0] & 1;
 }
 
-SECP256K1_INLINE static void secp256k1_fe_clear(secp256k1_fe *a) {
+SECP256K1_INLINE static void secp256k1_fe_set_zero(secp256k1_fe *a) {
     int i;
 #ifdef VERIFY
     a->magnitude = 0;
     a->normalized = 1;
+#endif
+    for (i=0; i<5; i++) {
+        a->n[i] = 0;
+    }
+}
+
+SECP256K1_INLINE static void secp256k1_fe_clear(secp256k1_fe *a) {
+    int i;
+#ifdef VERIFY
+    a->magnitude = 0;
+    a->normalized = 0;
 #endif
     for (i=0; i<5; i++) {
         a->n[i] = 0;