--- a/src/group_impl.h
+++ b/src/group_impl.h
@@ -195,9 +195,9 @@ static void secp256k1_ge_globalz_set_table_gej(size_t len, secp256k1_ge *r, secp
 
 static void secp256k1_gej_set_infinity(secp256k1_gej *r) {
     r->infinity = 1;
-    secp256k1_fe_clear(&r->x);
-    secp256k1_fe_clear(&r->y);
-    secp256k1_fe_clear(&r->z);
+    secp256k1_fe_set_zero(&r->x);
+    secp256k1_fe_set_zero(&r->y);
+    secp256k1_fe_set_zero(&r->z);
 }
 
 static void secp256k1_gej_clear(secp256k1_gej *r) {