--- a/src/group_impl.h
+++ b/src/group_impl.h
@@ -202,15 +202,15 @@ static void secp256k1_gej_set_infinity(secp256k1_gej *r) {
 
 static void secp256k1_gej_clear(secp256k1_gej *r) {
     r->infinity = 0;
-    secp256k1_fe_clear(&r->x);
-    secp256k1_fe_clear(&r->y);
-    secp256k1_fe_clear(&r->z);
+    SECP256K1_CLEANSE(r->x);
+    SECP256K1_CLEANSE(r->y);
+    SECP256K1_CLEANSE(r->z);
 }
 
 static void secp256k1_ge_clear(secp256k1_ge *r) {
     r->infinity = 0;
-    secp256k1_fe_clear(&r->x);
-    secp256k1_fe_clear(&r->y);
+    SECP256K1_CLEANSE(r->x);
+    SECP256K1_CLEANSE(r->y);
 }
 
 static int secp256k1_ge_set_xquad(secp256k1_ge *r, const secp256k1_fe *x) {