--- a/src/secp256k1.c
+++ b/src/secp256k1.c
@@ -376,9 +376,9 @@ int secp256k1_ecdsa_sign(const secp256k1_context* ctx, secp256k1_ecdsa_signature
             count++;
         }
         memset(nonce32, 0, 32);
-        secp256k1_scalar_clear(&msg);
-        secp256k1_scalar_clear(&non);
-        secp256k1_scalar_clear(&sec);
+        SECP256K1_CLEANSE(msg);
+        SECP256K1_CLEANSE(non);
+        SECP256K1_CLEANSE(sec);
     }
     if (ret) {
         secp256k1_ecdsa_signature_save(signature, &r, &s);
@@ -397,7 +397,7 @@ int secp256k1_ec_seckey_verify(const secp256k1_context* ctx, const unsigned char
 
     secp256k1_scalar_set_b32(&sec, seckey, &overflow);
     ret = !overflow && !secp256k1_scalar_is_zero(&sec);
-    secp256k1_scalar_clear(&sec);
+    SECP256K1_CLEANSE(sec);
     return ret;
 }
 
@@ -420,7 +420,7 @@ int secp256k1_ec_pubkey_create(const secp256k1_context* ctx, secp256k1_pubkey *p
         secp256k1_ge_set_gej(&p, &pj);
         secp256k1_pubkey_save(pubkey, &p);
     }
-    secp256k1_scalar_clear(&sec);
+    SECP256K1_CLEANSE(sec);
     return ret;
 }
 
@@ -469,8 +469,8 @@ int secp256k1_ec_privkey_tweak_add(const secp256k1_context* ctx, unsigned char *
         secp256k1_scalar_get_b32(seckey, &sec);
     }
 
-    secp256k1_scalar_clear(&sec);
-    secp256k1_scalar_clear(&term);
+    SECP256K1_CLEANSE(sec);
+    SECP256K1_CLEANSE(term);
     return ret;
 }
 
@@ -515,8 +515,8 @@ int secp256k1_ec_privkey_tweak_mul(const secp256k1_context* ctx, unsigned char *
         secp256k1_scalar_get_b32(seckey, &sec);
     }
 
-    secp256k1_scalar_clear(&sec);
-    secp256k1_scalar_clear(&factor);
+    SECP256K1_CLEANSE(sec);
+    SECP256K1_CLEANSE(factor);
     return ret;
 }