@sipa 2017-03-22
Overall concept ACK.

I'm not sure I like the separation of `secp256k1_fe_verify_init`, as it places a responsibility on the caller that they shouldn't have (as in: the group code ideally shouldn't need to know there is such a thing as debug variables in `secp256k1_fe` that need initialization). Given that `secp256k1_fe_clear` remains, can't the `secp256k1_fe_verify_init` logic remain in there? Then you can still use a strict separation between `SECP256K1_CLEANSE(fe)` (meaning `fe` won't be used anymore) and `secp256k1_fe_clear(&fe)` (meaning `fe` must get the value 0). You could even rename the latter to `secp256k1_fe_set_zero` perhaps.