@isle2983 2017-03-27
Great suggestions. I introduced `secp256k1_fe_set_zero()` and re-worked the commit history to avoid the wrangling of the `#define VERIFY` stuff by converting the remaining clients of `secp256k1_fe_set_zero()` over first. `secp256k1_fe_clear()` is then replaced with `SECP256K1_CLEANSE()` along with the others. `secp256k1_fe_set_zero()` remains and encapsulates the (former) `verify_init()` initialization logic.