@real-or-random 2019-05-23
Yes, that's my plan; consider this PR adopted by me. I just couldn't find time to work on this so far.