@sipa 2019-07-23
Closing as superseded by #636.